less repeated code!
